% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_reporting_tool_environment}
\alias{create_reporting_tool_environment}
\title{Create reporting tool environment}
\usage{
create_reporting_tool_environment()
}
\description{
This function creates the reporting tool environment used to
store the id values required as primary and foreign keys in
the reporting tool database
}
