% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_date}
\alias{format_date}
\title{Format single year date}
\usage{
format_date(date)
}
\arguments{
\item{date}{integer or character, 4 digits}
}
\value{
character, 4 digits with end dash
}
\description{
This function formats dates consisting of a 4 digit year, adding
a dash at the end so that the reporting tool can handle these single
dates in a year range. It also replaces the unicode dash character
}
\examples{
format_date(2018)
}
