% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_date_period}
\alias{format_date_period}
\title{Format date period}
\usage{
format_date_period(start_date, end_date)
}
\arguments{
\item{start_date}{4 digit integer, start date}

\item{end_date}{4 digit integer, end date}
}
\value{
character, yyyy- or yyyy-yyyy date period format
}
\description{
This function concatenates the start and end dates
into required yyyy- or yyyy-yyyy date period formats
}
\examples{
format_date_period(2013, 2018)
}
