% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_text}
\alias{format_text}
\title{Format text}
\usage{
format_text(text)
}
\arguments{
\item{text}{character, free text}
}
\value{
character, formatted free text
}
\description{
This function formats free text handling NA characters, end of line
characters, double quotes, URL characters and converting unicode
characters to the most appropriate ASCII character
}
\examples{

}
