% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\docType{data}
\name{information_scientific_name_look_up}
\alias{information_scientific_name_look_up}
\title{Scientific name information}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 322 rows and 5 columns.}
\usage{
information_scientific_name_look_up
}
\description{
A dataset a look up between the scientific names found in the
submitted 2013 species distribution maps shape file and Article 12
species codes. Scientific_name_match is used to indicate scientific
names with a unique species code or where there is more than one
species code.
}
\details{
#' @format A tibble with 332 rows and 5variables
\describe{
\item{speciescode}{bird species code}
\item{scientific_name}{bird scientific name}
\item{sub_unit}{article 12 checklist sub unit}
\item{scientific_name_match}{unique species code for scientific name or not}
\item{common_name}{bird common name}
}
}
\keyword{datasets}
