% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\docType{data}
\name{information_species_look_up}
\alias{information_species_look_up}
\title{Species information}
\format{A tibble with 553 rows and 4 variables
\describe{
\item{speciescode}{bird species code}
\item{speciesname}{scientific name}
\item{sub_unit}{bird sub_unit}
}}
\usage{
information_species_look_up
}
\description{
A dataset containing the Article 12 species information, taken from
the Birds Directive reporting portal. Checklist for bird species last
updated 2018-07-05
}
\keyword{datasets}
