% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_effectiveness_mp}
\alias{recode_effectiveness_mp}
\title{Effectiveness of management plans for huntable species in non-Secure status}
\usage{
recode_effectiveness_mp(effectiveness)
}
\arguments{
\item{effectiveness}{character, effectiveness text}
}
\value{
character, abbreviated effectiveness text
}
\description{
This function changes the effectiveness SAP text to
the abbreviated text used in the reporting tool
}
\examples{
recode_effectiveness_mp("further deteriorating") 
}
