% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_effectiveness_sap}
\alias{recode_effectiveness_sap}
\title{Effectiveness of SAPs}
\usage{
recode_effectiveness_sap(effectiveness)
}
\arguments{
\item{effectiveness}{character, effectiveness text}
}
\value{
character, abbreviated effectiveness text
}
\description{
This function changes the effectiveness SAP text to
the abbreviated text used in the reporting tool
}
\examples{
recode_effectiveness_sap("further deteriorating away from the plan’s aim/objective(s)")
}
