% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_measures_identified}
\alias{recode_measures_identified}
\title{Recode measures identified}
\usage{
recode_measures_identified(measure)
}
\arguments{
\item{measure}{character, measures identified text}
}
\value{
character, measures identified abbreviated text
}
\description{
This function changes the descriptive measures identified text
to the abbreviated text used in the reporting tool
}
\examples{
recode_measures_identified("Measures needed but cannot be identified")
}
