% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_measures_location}
\alias{recode_measures_location}
\title{Recode measures location}
\usage{
recode_measures_location(location)
}
\arguments{
\item{location}{character, measures location text}
}
\value{
character, measures location abbreviation text
}
\description{
This function changes the descriptive measures location text
to the abbreviated text used in the reporting tool
}
\examples{
recode_measures_location("Both inside and outside Natura 2000")
}
