% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_measures_purpose}
\alias{recode_measures_purpose}
\title{Recode species measures purpose}
\usage{
recode_measures_purpose(purpose)
}
\arguments{
\item{purpose}{character, measures purpose text}
}
\value{
character, measures purpose abbreviated text
}
\description{
This function changes the descriptive measures purpose text
for species to the abbreviated text used in the reporting tool
}
\examples{
recode_measures_purpose_species("Maintain the current distribution, population and/or habitat for the species")
}
