% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_measures_response}
\alias{recode_measures_response}
\title{Recode measures response}
\usage{
recode_measures_response(response)
}
\arguments{
\item{response}{character, measures response text}
}
\value{
character, measures response abbreviated text
}
\description{
This function changes the descriptive measures response text
to the abbreviated text used in the reporting tool
}
\examples{
recode_measures_response("Medium-term results (within the next two reporting periods, 2019-2030)")
}
