% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_methods_used}
\alias{recode_methods_used}
\title{Recode methods used}
\usage{
recode_methods_used(methods_used)
}
\arguments{
\item{methods_used}{character, method used text}
}
\value{
character, methods used abbreviated text
}
\description{
This function changes the descriptive methods used text to
the abbreviated text used in the reporting tool
}
\examples{
recode_methods_used("Based mainly on expert opinion with very limited data")
}
