% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_plans}
\alias{recode_plans}
\title{Recode plans}
\usage{
recode_plans(plans)
}
\arguments{
\item{plans}{character, plans text}
}
\value{
character, abbreviated plans text
}
\description{
This function changes the descriptive plans text to
the abbreviated text used in the reporting tool
}
\examples{
recode_plans("Brief Management Statement")
}
