% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_population_units}
\alias{recode_population_units}
\title{Recode population units}
\usage{
recode_population_units(population_unit)
}
\arguments{
\item{population_unit}{character, population unit text}
}
\value{
character, population unit abbreviated text
}
\description{
This function changes the descriptive population units text to
the abbreviated text used in the reporting tool
}
\examples{
recode_population_units("number of adults")
}
