% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_pressure_threat_locations}
\alias{recode_pressure_threat_locations}
\title{Recode pressure and threat locations}
\usage{
recode_pressure_threat_locations(location)
}
\arguments{
\item{location}{character, location text}
}
\value{
character, abbreviated location text
}
\description{
This function changes the descriptive location text
to the abbreviated text used in the reporting tool.
Includes both the numeric and text location options
}
\examples{
#' recode_ranking("inside the Member State")
}
