% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_ranking}
\alias{recode_ranking}
\title{Recode ranking}
\usage{
recode_ranking(ranking)
}
\arguments{
\item{ranking}{character, ranking text}
}
\value{
character, abbreviated ranking text
}
\description{
This function changes the descriptive ranking text
to the abbreviated text used in the reporting tool
}
\examples{
recode_ranking("high importance")
}
