% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_reason_change}
\alias{recode_reason_change}
\title{Recode reason change}
\usage{
recode_reason_change(reason)
}
\arguments{
\item{reason}{character, reason change text}
}
\value{
character, abbreviated freason change text
}
\description{
This function changes the descriptive reason change
text to the abbreviated text used in the reporting tool
}
\examples{
recode_reason_change("Improved knowledge/more accurate data")
}
