% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_season}
\alias{recode_season}
\title{Recode season}
\usage{
recode_season(season)
}
\arguments{
\item{season}{character, season text}
}
\value{
character, season abbreviated text
}
\description{
This function changes the season text to
the abbreviated text used in the reporting tool
}
\examples{
recode_season("Breeding")
}
