% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_trends}
\alias{recode_trends}
\title{Recode trends}
\usage{
recode_trends(trends)
}
\arguments{
\item{trends}{character, trends text}
}
\value{
character, abbreviated trends text
}
\description{
This function changes the descriptive trends text to
the abbreviated text used in the reporting tool
}
\examples{
recode_trends("Increasing (+)")
}
