% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_type_of_estimate}
\alias{recode_type_of_estimate}
\title{Recode type of estimate}
\usage{
recode_type_of_estimate(estimate)
}
\arguments{
\item{estimate}{character, type of estimate text}
}
\value{
character, type of estimate abbreviated text
}
\description{
This function changes the descriptive type of estimate
text to the abbreviated text used in the reporting tool
}
\examples{
recode_type_of_estimate("Multi-year mean")
}
