% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{reverse_yes_no}
\alias{reverse_yes_no}
\title{Reverse yes and no}
\usage{
reverse_yes_no(yes_no_blank)
}
\arguments{
\item{yes_no}{character, yes, no or NA_character_}
}
\value{
character, yes and no are swopped around, blank converted to yes
}
\description{
This function reverses yes and no as species categories 5.11, 6.16
and 11.7 in data capture spreadsheet ask for change (yes or no) where
as in reporting tool the same categories ask for no change (yes or no)
}
\details{
Blanks are converted to yes (no change) in the reporting tool as the
reporting tool only stores two states yes or no
}
\examples{
reverse_yes_no(c("Yes", "No", NA))
}
