% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\docType{data}
\name{information_common_name_look_up}
\alias{information_common_name_look_up}
\title{Common name information}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 334 rows and 6 columns.}
\usage{
information_common_name_look_up
}
\description{
A dataset a look up between the common names found in the
Article 12 spreadsheets and Article 12 species codes.
Match is used to indicate common names with a unique species
code or where there is more than one species code. For these
common names additional information is provided to allow a
manual selection of the correct species code.
}
\details{
#' @format A tibble with 334 rows and 6 variables
\describe{
\item{speciescode}{bird species code}
\item{common_name}{bird common name}
\item{additional_info}{additional name information}
\item{common_name_match}{unique species code for common name or not}
\item{scientific_name}{article 12 checklist scientific name}
\item{sub_unit}{article 12 checklist sub unit}
}
}
\keyword{datasets}
